Join type definitions into a single file
Makes it easier to glance at the whole submodule without switching files.
