--- plugins/gatsby-source-cd-airtable/src/errors/connection-error.ts
@@ -1,5 +1,0 @@
-export class ConnectionError extends Error {
-  constructor(message?: string) {
-    super(message)
-  }
-}
