--- plugins/gatsby-source-cd-airtable/src/interfaces/airtable-partner.ts
@@ -1,9 +1,0 @@
-import { AirTableRecord } from './airtable-record'
-
-export interface AirTablePartner extends AirTableRecord {
-  fields: {
-    name: string
-    url: string
-    logoUrl: string
-  }
-}
