--- plugins/gatsby-source-cd-airtable/src/interfaces/airtable-record.ts
@@ -1,4 +1,0 @@
-export interface AirTableRecord {
-  id: string
-  fields: Record<string, unknown>
-}
