--- plugins/gatsby-source-cd-airtable/src/interfaces/index.ts
@@ -1,5 +1,0 @@
-export * from './airtable-partner'
-export * from './airtable-project'
-export * from './partner'
-export * from './plugin-options'
-export * from './project'
