--- plugins/gatsby-source-cd-airtable/src/interfaces/partner.ts
@@ -1,7 +1,0 @@
-import { SourceNode } from './source-node'
-
-export interface Partner extends SourceNode {
-  name: string
-  url: string
-  logoUrl: string
-}
