--- plugins/gatsby-source-cd-airtable/src/interfaces/plugin-options.ts
@@ -1,6 +1,0 @@
-export interface PluginOptions {
-  projectsTableName: string
-  tagsTableName: string
-  volunteersTableName: string
-  partnersTableName: string
-}
