--- plugins/gatsby-source-cd-airtable/src/interfaces/source-node.ts
@@ -1,3 +1,0 @@
-export interface SourceNode {
-  rowId: string
-}
