+++ plugins/gatsby-source-cd-airtable/src/mocks/mock-partners.ts
@@ -1,4 +1,4 @@
-import { Partner } from '../interfaces'
+import { Partner } from '../types'
 
 export const getMockPartners: () => Partner[] = () => [
   {
