+++ plugins/gatsby-source-cd-airtable/src/mocks/mock-volunteers.ts
@@ -1,4 +1,4 @@
-import { Volunteer } from '../interfaces'
+import { Volunteer } from '../types'
 
 export const getMockVolunteers: () => Volunteer[] = () => [
   {
