+++ plugins/gatsby-source-cd-airtable/tests/transformers.spec.ts
@@ -4,7 +4,7 @@ import {
   transformTags,
   transformVolunteers,
 } from '../src/transformers'
-import { AirTableProject } from '../src/interfaces'
+import { AirTableProject } from '../src/types'
 
 describe('transformers', () => {
   describe('transformProjects', () => {
